requirements.txt
pyttsx3
speechRecognition
wikipedia
gradio_client
PyAudio
flet
numpy
scipy
clap-detector
python-dotenv
psutil
sarvamai